+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Send parameters in request body instead of query string for query based services and EC2
 - Allow AWS credentials in environment variables to have a custom prefix
 - Fix bug in presigned URLs for S3
+- Add `From<Result<T, E>> for RusotoFuture<T, E>` implementation for mocking.
 
 ## [0.33.1] - 2018-08-07